setup upload and delete old files-endpoint files
