+++ gateway/api/v1/urls.py
@@ -46,9 +46,6 @@ router.register(
     basename=v1_views.JobViewSet.BASE_NAME,
 )
 router.register(
-    r"files", v1_views.FilesViewSet, basename=v1_views.FilesViewSet.BASE_NAME
-)
-router.register(
     r"catalog",
     v1_views.CatalogViewSet,
     basename=v1_views.CatalogViewSet.BASE_NAME,
