+++ gateway/api/v1/views/__init__.py
@@ -5,6 +5,5 @@ Specification for V1 version of the API.
 """
 
 from .catalog import CatalogViewSet
-from .files import FilesViewSet
 from .jobs import JobViewSet
 from .programs import ProgramViewSet
