+++ gateway/api/v1/views/files/delete.py
