+++ gateway/api/views/__init__.py
@@ -5,6 +5,5 @@ Version views inherit from the different views.
 """
 
 from .catalog import CatalogViewSet
-from .files import FilesViewSet
 from .jobs import JobViewSet
 from .programs import ProgramViewSet
